\name{wica.bdy}
\alias{wica.bdy}
\docType{data}
\title{
National Park Boundary
}
\description{
A sample object of class "\code{SpatialPolygonsDataFrame}"\{\pkg{sp}\} representing the boundary of Wind Cave National Park
}
\format{
Formal class "\code{SpatialPolygonsDataFrame}" [package "sp"] with 5 slots
}
\usage{data(wica.bdy)}
\source{
Prepared in October 2010 from an ESRI shapefile provided by Kevin Kovacs, NPS Wind Cave National Park
}
\examples{
data(wica.bdy)
str(wica.bdy)
plot(wica.bdy)
}
\keyword{datasets}
