\name{wica.gpc}
\alias{wica.gpc}
\alias{thro.gpc}
\docType{data}
\title{
National Park Boundary
}
\description{
An object of class \code{gpc.poly} for use in polygon clipping. The class is provided by package \pkg{gpclib}.
}
\usage{data(wica.gpc)}
\source{
Derived from \code{wica.bdy}.
}
\examples{
data(wica.gpc)
plot(wica.gpc)
}
\keyword{datasets}
