\name{wild1-package}
\alias{wild1-package}
\alias{wild1}
\docType{package}
\title{
R tools for wildlife research and management
}
\description{
Provides classes, methods, and examples for use in wildlife research and management.
}
\details{
\tabular{ll}{
Package: \tab wild1\cr
Type: \tab Package\cr
Version: \tab 1.08\cr
Date: \tab 2010-12-03\cr
License: \tab GPL-3 \cr
LazyLoad: \tab yes\cr
}
}
\author{
Glen A. Sargeant

U.S. Geological Survey

Maintainer: \email{glen_sargeant@usgs.gov}
}
\section{Required disclaimers}{
Although this program has been used by the USGS, no warranty, expressed or implied, is made by the USGS or the United States Government as to the accuracy and functioning of the program and related program material nor shall the fact of distribution constitute any such warranty, and no responsibility is assumed by the USGS in connection therewith.  Any use of trade, product, or firm names is for descriptive purposes only and does not imply endorsement by the U.S. Government.}
\keyword{ package }
\keyword{ survival }
\keyword{ manip }
\keyword{ aplot }
\keyword{ hplot }

