\name{wica.gpc}
\alias{wica.gpc}
\alias{wica.hole.gpc}
\docType{data}
\title{
National Park Boundary
}
\description{
A sample object of class \code{gpc.poly} for use in polygon clipping. The class is provided by package \pkg{gpclib}, which is subject to a restricted license and is not free for commercial use.
}
\usage{data(wica.gpc)}
\source{
Derived from \code{wica.bdy}.  For purposes of illustration, \code{wica.hole.gpc} includes a fictitious "hole."
}
\examples{
data(wica.gpc)
plot(wica.gpc)
}
\keyword{datasets}
